Optimize latest album/single fetch
By limiting the market, the Spotify API always returns the latest album
or single in the Artist album's endpoint. With this information at hand,
we can limit the number of calls to the API, which greatly increases the
time it takes to fetch all the information.

First, we call the Artist album's api to fetch the latest "album".
Later, we call it again, but this time to fetch the latest "single".
With both IDs, we can then call the album show endpoint, passing in the
IDs (the api permits that).

This limits the number of calls to only 3, which is not perfect, but it
is a great improvement from the previous implementation.