Add User POST api endpoint
This "barebone" version allows _anyone_ to create an User using the API.
The route is: POST /api/v1/user

Since this is a JSON endpoint, it expects a json body in order to create
the resource:

{"name": "Matias", "uid": "1241", "provider": "spotify", "email":
"lla@bar.com", "metadata": "{\"foo\": \"bar\"}"}

Notice the "metadata" key is a string containing a JSON object. It is
important to escape the `"` keys, otherwise it will generate an invalid
JSON object.