Remove Grape in favor of API implementation "by hand"
Grape is great, but it add a level of abstraction that I believe makes
things harder instead of easier. That's why I'm dropping it.

To replace it, I'm doing the good old rails routes and namespaces, which
is more familiar.

To serialize data, I added ActiveModelSerializers, which seems to be a
very good solution. It also complies with the json:api specification
(http://jsonapi.org/).