Add `/api/v1/me/releases` endpoint
Endpoint returns the current user's releases based on artists that he or
she follows.

For now this will return all latest 9 releases, without any sort for
state checking. In the future we might want to check exactly what we
should be returning here, to avoid unnecessary calls and long responses.