Add spotify tokens swap/refresh endpoints
This is needed in order to the iOS app get a token, which will later be
used to make spotify calls on behalf of the user.

This implementation was fully copied from
https://github.com/simontaen/SpotifyTokenSwap/ and adapted to use our
own environment variables.

The endpoints are:

POST /spotify/tokens/swap (takes only a `code` param)
POST /spotify/tokens/refresh (takes only a `refresh_token` param)