Add `api/v1/users/find` endpoint
New endpoint should be used to find user using his or her email. If
there's an user with the provided email, the endpoint will return the
serialized user. If not, it returns nil.

This is helpful for clients that want to check if an user exists before
hitting the create endpoint.