Use JWT to authorize the Profile API endpoint
Requests to the API should have the following encoded JSON inside the
"Authorization: Bearer <token>" header:

```
{
  "claims": {
    "email" : "matiasleidemer@gmail.com"
  }
}
```

With this, the old X-Released-User-Token doesn't work anymore. Since the
APP is not up and running yet, it is OK to break the API like we're
doing here, but in the near future, it might me wise to keep the old API
working while clients migrate to the new one.

closes #23