Include yesterday when checking for send new notifications
Sometimes spotify makes some releases available *after* our job runs.
This means that users won't get their new releases when they get
released. I've already pushed the time when the job runs a little
further (15:00 UTC instead of 12:00 UTC), but we can't be 100% sure if
that will do the trick.

In order to deliver those releases, I've changed the
Album#released_today? method to include yesterday in the check as well.

This means we might send yesterday's releases to the user, which is
better than not sending them at all. I still doesn't know exactly how to
properly address this issue, but this seems like a good alternative
until then.